+++ aluno.h
@@ -2,21 +2,21 @@
 struct tAlunoEntrada {
 	char nome[CARAC];
 	int RA;
-	float p1;
-	float p2;
-	float pt;
-	float po;
-	float media;
-    int situacao;
+	double p1;
+	double p2;
+	double pt;
+	double po;
+	double media;
+	int situacao;
 };
 
-void addAluno(tAlunoEntrada *vetAlunos, int &n, char *mod, int RA,float p1, float p2, float pt, float po);
+void addAluno(tAlunoEntrada *vetAlunos, int &n, char *mod, int RA,double p1, double p2, double pt, double po);
 void criaAprovados(int &n, tAlunoEntrada *vetAlunos);
 void criaReprovados(int &n, tAlunoEntrada *vetAlunos);
 void abreArquivo(tAlunoEntrada *vetAlunos, int &n);
 void buscaAluno(int &n, tAlunoEntrada *vetAlunos, char *nomedesejado);
-bool aprovador(float p1, float p2, float pt);
-void optativa(float p1, float p2, float pt, float po, float* r1, float* r2, float* r3);
+bool aprovador(double p1, double p2, double pt);
+void optativa(double p1, double p2, double pt, double po, double* r1, double* r2, double* r3);
 void ordenaAlunos(int &n, tAlunoEntrada *vetAlunos);
 void criaAtual(int &n, tAlunoEntrada *vetAlunos);
 void copiar(int &n, tAlunoEntrada *vetAlunosOriginal,  tAlunoEntrada *vetAlunosCopiado);
